==> 1_scripts/01_ROV_initial_wrangle.Rmd <==
---
title: "ROV_initial_wrangle"
author: "Gemma Galbraith"
date: "20/05/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This script is designed to take pre-prepared eventmeasure csv files (or similar) and pull out separate data frames for different analyses. Note that some manipulation in excel is required for the raw eventmeasure output e.g. removing the top three rows and the columns spacing out file data from observations. 

Fish
Trash
Sea Cucumbers
Other observations

In hindsight, it would have been good not to pre-process so much in excel. We lost the observation period/video times which would have been useful. Will proabbly come back to this and re do.


# Housekeeping
```{r}
rm(list=ls())
#dev.off()
#getwd()
```

#Packages
```{r}
library(readxl) #for data import if not in CSV format
library(tidyverse) #for tidying and organizing data
```

# Initial import
```{r}
# We need the reef, site name as a new variable to identify the source of the data
# Pulls each file in, merges and creates a filename variable. 

path <- "../0_data/ROV_data"

files <- list.files(path, pattern="csv", full.names=TRUE) %>%
    set_names()

data <- files %>% map_dfr(read_csv, .id="filename")

```

Notes on these data:

Would be good to keep updating the EventMeasure species file to a) avoid big wrangling tasks for fish names entered in the comments variable and b) avoid spelling mistakes. Could also add all the levels of the trash variable to the species file to allow that to be selected for easily.

Also, consider keeping the period/video time variable from the raw EventMeasure .csv file. This would make going back through videos much quicker, esp for things like sea cucumbers.

# Explore NA values for Transect
```{r}

which(is.na(data$Period))
data %>% filter(is.na(Period))
# The two still showing are some observations of cool stuff.

# Many NA values - requires going into the csv files and checking what's going on. May even require looking at the videos. Some are genuine observations made outside of any particular transect but these shouldn't be included in the statistical analysis if they we not record on a transect.

# NA values typically corrected directly in csv files
```


# Variable selection and creation
Would like to select the site name from the file name variable using regex eventually. For now, string subsetting does the job.

```{r}
#Tidy up to remove unwanted variables and create a site name variable, Reef variable and a survey date.
# NB. extra survey dates will be added to the case_when code line as and when more videos from other survey trips are added to data.

unique(data$OpCode) # shows that some OpCodes have not been named correctly. Use 'filename' to apply SURVEY and maybe REEF. Edna and Juliette Cay require a bit of special attention using the OpCode variable. We also need to specify whether the survey was stereo or not. Refer to ROV transect full summary csv.

merged <- data %>% mutate(
  Site=str_sub(filename,9,-13),
  Site=substring(Site,12),
  Site = case_when(OpCode == "Edna_1_JUL2021" ~ "EdnaCay_1",
                   OpCode == "Edna_2_JUL2021" ~ "EdnaCay_2",
                   OpCode == "Edna_3_JUL2021" ~ "EdnaCay_3",
                   OpCode == "Juliette1_July2021" ~ "JulCay_1",
                   TRUE ~ Site),
  Video = case_when(str_detect(Site, c("Edna|Jul|HerSurp")) ~ "Single",
          TRUE ~ "Stereo"),        
  Transect=as.factor(Period),
  Length_mm=as.double(Length_mm),
  Survey=as.factor(case_when(grepl("FEB2021", filename) ~ "Feb_21",
                             grepl("FEB2022", filename) ~ "Feb_22",
                             grepl("JUL2021", filename)~ "Jul_21")),
  Site=as.factor(Site))

# Edna is three sites that have been analysed in one eventmeasure file - some manual edits done to the csv but still needs some attention here. Use OpCode which has been manual edited to reflect the three different sites.

# Get Reef names

merged <- merged %>% mutate ( 
  Reef = as.factor(case_when(grepl("Li", OpCode) ~ "Lihou",
                   grepl("Flin", OpCode) ~ "Flinders",
                   grepl("Bou", OpCode) ~ "Bougainville",
                   grepl("Osp", OpCode) ~ "Osprey",
                   grepl("Chil", OpCode) ~ "Chillcot",
                   grepl("Edna", OpCode) ~ "Lihou",
                   grepl("Jul", OpCode) ~ "Lihou",
                   grepl("Hera", OpCode) ~ "Herald",
                   grepl("Holm", OpCode) ~ "Holmes",
                   grepl("Wil", OpCode) ~ "Willis",
                   grepl("Fred", OpCode) ~ "Frederick",
                   grepl("Ken", OpCode) ~ "Kenn",
                   grepl("Wrec", OpCode) ~ "Wreck",
                   grepl("Sau", OpCode) ~ "Saumarez",
                   grepl("HerS", OpCode) ~ "Herald's Surprise",
                   grepl("Mar", OpCode) ~ "Marion",)))

new.data<-select(merged,c(Survey,Reef,Site,Transect,Length_mm,Family,Genus,Species, Number,Comment, Video))
```



# Unique transect ID 
```{r}

new.data <- new.data %>% mutate(
  #Transect=str_sub(Transect,2), # if you want to get rid of "T"
  T_ID = paste(Survey,Site,Transect, sep = '_'))

```


# Separate fish, trash, cucumbers and cool stuff
At this stage we need to pull out the fish observations from other notable things spotted on the transects. Create new data frames for each thing.

For fish this is easy as only fish observations will at least have "Family", "Genus" and/or "Species". Use this as a filter. NB making sure you at least ID a fish to family level is key.

Fish
```{r, fish} 

fish <- new.data %>% filter(
  !is.na(Family))

write.csv(fish, file= "../0_data/ROV_fish_data.csv", row.names = FALSE)

```


Cucumbers
```{r}

cucs<- new.data %>% filter(
  grepl('cucumber|umber', Comment))%>% 
  select(- Family, - Genus,- Species)
  
  
write.csv(cucs, file= "../0_data/ROV_cucumber_data.csv", row.names = FALSE)

# Alternatively a partial match for the word cucumber if there may be variations of this spelling in the data

# cucs<- data %>% filter(
# str_detect(Comment, "cumber"))%>% 
# select(- Family, - Genus,- Species)
  
  

```

Trash
Selecting the trash is a bit more difficult, we need to know the levels to match the strings. It might be best to just get a data frame with no fish and no cucumbers so we can see what we have. Use the same is.na rule for "Family" but this time, those NA rows are what we want. Just drop the ! logical operator.

```{r}
trash <- new.data %>% filter(
  is.na(Family) & is.na(Species) & !str_detect(Comment, "cumber"))%>% 
  select(-Family, -Genus, -Species)

# Learning point: str_detect is very regex based in its searches. It also has a "negate" argument which will NOT search for the pattern you supply, works like the ! operator. e.g.str_detect(Comment, "cumber", negate = TRUE)

glimpse(trash)
trash$Comment
distinct(trash, Comment)

# We see from the above filtered data that other things from the comment variable have been selected. Exclude these by filtering only the known trash categories.

# "fishing_line" and "trash" are our only marine debris cats so far. Check for multiple spellings e.f. "Fishing Line" by including partial string match.

trash <- trash %>% filter(
  str_detect(Comment, "ishing") | str_detect(Comment, "trash" | str_detect(Comment, "Fishing Line")))
  # select(-Length_mm, -Number)

write.csv(trash, file= "../0_data/ROV_trash_data.csv", row.names = FALSE)


```

General/cool observations
Because the comments feature is the only flexible option for noting other observations in the videos, any other things noted in the transect shows up here. Use a similar filtering approach as above to isolate these observations.

```{r, coolstuff}
coolstuff <-data %>% filter(
  is.na(Family) & is.na(Species) & !str_detect(Comment, c("cumber|fishing_line|trash|Fishing Line|Fishing line")))%>% 
  select(-Family, -Genus, -Species)

coolstuff

# Check that there's nothing in here that shouldn't be in the other subsets


```
