==> 1_scripts/02_BRUV_fish_wrangle.Rmd <==
---
title: "02_BRUV_fish_wrangle"
output: html_document
date: '2022-06-22'
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This markdown takes the prepared fish data from the initial BRUV wrangle (01_BRUV_initial_wrangle). Is removes unnecessary variables, forms a unique transect/drop ID and merged with the site information.

#Housekeeping
```{}
rm(list=ls())
#dev.off()
getwd()
```

#Libraries
```{r}
library(tidyverse)
library(readr)
```

# Data import

```{r}
fish <-read_csv("../0_data/BRUV_fish_data.csv")

fish<-fish %>% select(Survey,Reef, Site,Drop,Frame,Depth, Comment, Family, Genus, Species,Code, MaxN, Species_name)

fish <- fish %>% mutate(
  Reef=as.factor(Reef),
  Survey=as.factor(Survey),
  Site=as.factor(Site),
  Drop =as.character(Drop),
  Depth = as.factor(Depth),
  Species_name = as.factor(Species_name))
```

Create Unique Transect(Drop) ID, T_ID - must be T to match potential future analyses with ROV data.

```{r}
# Convert Drop1 to D1 throughout
fish<-fish%>%mutate(Drop=as.factor(str_remove_all(Drop, "[DROP]")))

fish<-fish%>%mutate(Drop=paste("D",Drop, sep=""))


fish<-fish %>% mutate(
  T_ID = as.factor(paste(Survey,Site,Drop, sep = "_"))
)

```

# Import drop locations summary info
```{r}
sites <-read_csv("../0_data/BRUV_site_drop_summary.csv")



```
# Define variables and create matching T_ID
```{r}
sites <- sites %>% mutate(
  Reef=as.factor(Reef),
  Survey=as.factor(Survey),
  Site=as.factor(Site),
  Drop =as.factor(paste("D",Drop, sep ="")),
  Depth = as.factor(Depth),
  Habitat = as.factor(Habitat),
  T_ID = as.factor(paste(Survey,R_site_code,Drop, sep = "_")),
  Site = as.factor(R_site_code)) %>% 
  select(-R_site_code)

levels(sites$Site)
levels(sites$Depth)
```

```{r}
levels(fish$T_ID)
levels(fish$Site)
levels(fish$Depth)
#check matching TID - ALL OF THESE SHOULS RETURN ZERO ANTI-MATCHES
anti_join(fish,sites, by ="T_ID")
anti_join(fish,sites, by ="Survey")
anti_join(fish,sites, by ="Site")
anti_join(fish,sites, by ="Reef")
anti_join(fish,sites, by ="Drop")
anti_join(fish,sites, by ="Depth")

# Drops that were made but were not suitable for anlaysis /failed deployments - check against video analyses log
anti_join(sites,fish, by ="T_ID")
```


#Merge
```{r}
merged.data<-left_join(fish, sites, by = c("Survey","Reef","Site","Drop","T_ID"))

merged.data%>%filter(is.na(Lat))
is.na(merged.data$Habitat)

# Depth discrepancies?
depths<-merged.data%>%select(T_ID,Depth.x, Depth.y)%>%
  mutate(Depth.x = as.numeric(Depth.x),
         Depth.y = as.numeric(Depth.y))

depths2<-subset(fish$Depth, sites$Depth)

mismatches = subset(depths, depths$Depth.x != depths$Depth.y)

# Two sites incorrect in BRUV csv files - checked survey logs
fish$Depth[fish$T_ID=="Feb_21_Flinders_S2_D3"] <-55
fish$Depth[fish$T_ID=="Oct_21_Ashmore_2_D9"] <-37

# Merge Again
merged.data<-left_join(fish, sites, by = c("Survey","Reef","Site","Drop","Depth","T_ID"))

merged.data%>%filter(is.na(Lat)) # all obs now match by the variables above
merged.data%>%filter(is.na(Habitat))


```
# Missing drops - proabbly not usbale drops
```{r}
missing<-anti_join(sites,fish, by = "T_ID")

```


# Write csv
```{r}
write.csv(merged.data, file= "../2_outputs/BRUV_cleaned_maxn_sites.csv", row.names = FALSE)


```
