==> 1_scripts/02_ROV_fish_wrangle.Rmd <==
---
title: "ROV_fish_wrangle"
author: "Gemma Galbraith"
date: "26/05/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#HouseKeeping
```{R}
rm(list=ls())
#dev.off()
#getwd()

```

#load libraries
```{r}
library(tidyverse)
library(rfishbase)
library(lubridate)
```

#Import fish data
```{r}
fish <-read_csv("../0_data/ROV_fish_data.csv")

fish <- fish %>% mutate(
  Transect=as.factor(Transect),
  Reef=as.factor(Reef),
  T_ID=as.factor(T_ID),
  Length_mm=as.double(Length_mm),
  Survey=as.factor(Survey),
  Video = as.factor(Video),
  Site=as.factor(Site))

```

#Transect infomation csv.
Take the output file "transect_summary_data" and merge by TID. This now assigns the details e.g. depth, temp, complexity etc to each observation

```{r}
transect_data <-read_csv("../0_data/ROV_Transect_Summary_Analysis.csv")

transect_data<-select(transect_data,c(-Day_no, -Dive_no, -Trash, -Cucumbers,- Notes))

transect_data$Transect<-sub("^", "T", transect_data$Transect)

# Redefine some variables
transect_data <- transect_data %>% mutate(
  Reef=as.factor(Reef),
  Site=as.factor(Site),
  Transect=as.factor(Transect),
  Date=as.Date(Date, format = "%d/%m/%Y"),
  Habitat=as.factor(Habitat),
  Aspect=as.factor(Aspect),
  Survey=as.factor(case_when(month(Date) ==2 & year(Date) == 2021 ~ "Feb_21",
                             month(Date) ==2 & year(Date) == 2022 ~ "Feb_22",
                             month(Date) == 7 & year(Date) == 2021 ~ "Jul_21")),
  Site=as.factor(Site),
  T_ID = paste(Survey,Site,Transect, sep = '_'),
  T_ID=as.factor(T_ID))

```

#Merge fish and transect data - prechecks for matching transects
```{r}

#  check you have correct merge by unique values for T_ID - this is critical
distinct(transect_data, T_ID) # 163 transects
distinct(fish, T_ID) #156 rows - some transects had no fish so this is fine for now
# Check fish observations that don't have a transect number - 0 all good
which(is.na(fish$Transect))

fish %>% filter(is.na(Transect))
fish %>% filter(is.na(T_ID))

# There were some NA values - I had to manually go back through the csv files and videos to check.

levels(fish$T_ID)
unique(fish$T_ID) #155 unique transect IDs in the fish
unique(transect_data$T_ID) #163 transects in the transect_data
# This means we have 7 transects analysed which had no fish, could be lagoons ect
# Needs to be checked, do this using an anti-join - will select all rows from df1 (transect_data) that are not present in df2 (fish)

transect_data %>% anti_join(fish, by = "T_ID")
fish %>% anti_join(transect_data, by = "T_ID")
levels(transect_data$T_ID)
levels(fish$T_ID)

# 26/05/22 Saumarez, Wreck, Kenn, Marion and Frederick data are not yet in the project. Others in lagoon habitat are likely true zero observation transects. This should be made note of in the transect summary as the video anlaysis is done.

# Discrepancy between fish Site BVille and transect Site Bville. Changed in Transect summary data csv.
# Chilcott was spelt incorrectly in the Eventmeasure file but correctly in the transect_data file, consequently wasn't matching - changed manually in csv file now.
```

#The Merge
```{r}

# data <- merge(fish, transect_data, by =c("T_ID"))
# Check that you retain all observations from fish in the merge. I.e "data" should have the same no of obs as "fish" just with the info from "transect_data" appended.

data<- fish %>% 
  left_join(transect_data,by ="T_ID") %>% 
  select(-Reef.y, -Site.y, -Transect.y, -Survey.y) %>% 
  rename(Reef = Reef.x, Survey = Survey.x, Site = Site.x, Transect = Transect.x)

# Removed and rename duplicate variables created in the merge

```

# Missing species

If species names have been added in the "comments" variable in EventMeasure, these need to be pulled into the family, genus , species variables. Some comments have average length estimates for that species appended to the species name. Using regex and or coalesce we can split anything numeric into another variable and eventually transfer to the length_mm variable.

```{r}
# Inspect the comments variable and identify values 
data %>% filter(
  !is.na(Comment)
)

# Remove any observation with a comment value from the main data and create a sep df. Then just work on data2.
data2<-data %>% filter(
  !is.na(Comment)
)

data<-data %>% filter(
  is.na(Comment)
)

```

# Main data still needs to be checked - may have missing species
```{r}
# Inspect the species variable and identify values 
data %>% filter(
  is.na(Species)
)

data3<-data %>% filter(
  is.na(Species)
)

data<-data %>% filter(
  !is.na(Species)
)

```

```{r} 
# Fix data3

data3

# Paracentropyge multifasciata was named incorrectly
data3$Genus[2] <- "Paracentropyge"
data3$Species[2] <- "multifasciata"
data3$Family[2] <- "Pomacentridae"

data3
# everything else here just needs "sp" adding to the species column
# All NA species get "sp"
data3$Species[is.na(data3$Species)]<- "sp"
```

# wokring on subset data2
```{r}

# Most follow the "species_length.est" pattern but there are some funky entries. Split data2 again.
# Fish species and species with average lengths follow the pattern "species_length". Use this in regex to select
# /[a-z]+\_[0-9]+\.\d|[a-z]+\_[0-9]+/mg - worked in 101 but not here...
# funky_data <- data2 %>% filter(!str_detect(Comment, pattern = "/^\\[a-z]+\\_[0-9]+\\.\\d$|^\\[a-z]+\\_[0-9]+$"))

funky_data <- data2 %>% filter(!str_detect(Comment, pattern = "^[a-z]+\\_[0-9]+|^[0-9]+"))
# funky data is everything else. #

data2 <- data2 %>% filter(str_detect(Comment, pattern = "^[a-z]+\\_[0-9]+|^[0-9]+"))
# data2 now has all the "species_length.est" observations 

single.species <- funky_data %>% filter (str_detect(Comment, pattern = "^[a-z]+$"))
funky_data <- funky_data %>% filter (!str_detect(Comment, pattern = "^[a-z]+$"))

```

# data2 - deal with the "species_length.est" value in comments
```{r}
# First split the comment by the "_"
data2<- data2 %>% separate(
  Comment, c("temp.species","temp.length"),sep = "_", remove = FALSE
) %>% mutate(
  temp.length=as.double(temp.length)
)


# Missing pieces here are a random depth observation for H.marcosi and a funky trout.

# NOTE these length ests are in cm - converrt to mm to keep consistent with EM estimates! Will stuff uop your biomass later on.

# temp.lengths are all in cm - convert to mm
data2<- data2 %>% mutate(
  temp.length = temp.length*10)

data2 %>% filter(
  is.na(temp.length)
)

# Random missing length for a trout - check video
data2$Length_mm[117] <- 200
data2$temp.species[117] <- "leopardus"
data2$temp.length[117] <- 200

# COALESCE temp.species and temp.length to fill NA values
data2 <- data2 %>% mutate(
  Species=coalesce(Species, temp.species),
  Length_mm=coalesce(Length_mm, temp.length))

data2<- select(data2, c(-temp.species, -temp.length))
```


# Single species names
```{r}
# some comments are just single species names that need to be moved to the species column.
# Then ready to R bind with main data
single.species <- single.species %>% mutate(
  Species=coalesce(Species, Comment))
```

# Dealing with the funky data

This will be slightly different each time you add new survey data to the script. You will need to work through case by case.
https://stackoverflow.com/questions/70868379/replace-one-column-with-another-using-regex-matching-in-r

```{r}

funky_data

funky_data$Genus[38] <- "Paracentropyge"
funky_data$Species[38] <- "multifasciata"
funky_data$Length_mm[38] <- 103
funky_data$Length_mm[34] <- 115


funky_data<-funky_data %>% mutate(
  Genus = case_when(Comment == "Neoniphon sammara" ~ "Neoniphon",
                    Comment == "Neoniphon opercularis" ~ "Neoniphon",
                    TRUE ~ Genus),
  Species = case_when(Comment == "Neoniphon sammara" ~ "sammara",
                      Comment == "Neoniphon opercularis" ~ "opercularis",
                      Comment == "ALBIPECTORALIS" ~ "albipectoralis",
  TRUE ~ Species))


funky_data

```

# Row bind corrected single species and funky data back to main data
```{r}
fin.data<-rbind(data,data2,data3,single.species,funky_data)
# Check that your final data here has the same observations as "fish" df that you started with. Otherwise you have lost some in the cleaning processes above :?

# All NA species get "sp"
fin.data$Species[is.na(fin.data$Species)]<- "sp"
# spp <- sp in Species
fin.data$Species[fin.data$Species == "spp"] <- "sp"
```


```{r}

write.csv(fin.data, file= "../2_outputs/ROV_cleaned_fish.csv", row.names = FALSE)

```
