==> 1_scripts/03_ROV_fish_biomass_constants.Rmd <==
---
title: "Apply Genus or Family level Biomass a and b constants"
author: "Gemma Galbraith"
date: "30/05/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

After initial cleaning and wrangle, some observations remain at family or genus level. Apply the fish-families or fish-genera data from MERMAID to assign approximate a and b values for these species.

#HouseKeeping
```{R}
rm(list=ls())
#dev.off()
#getwd()

```

#load libraries
```{r}
library(tidyverse)
library(rfishbase)
```

#Import fish data
This uses the Final_fish data created by the scripts "ROV_initial_wrangle" to isolate just fish observations and then "ROV_fish_wrangle" which identifies and corrects errors in species names, has assigned a and b values where species are identified etc.

```{r}
fish <-read_csv("../2_outputs/ROV_cleaned_fish.csv")

fish <- fish %>% mutate(
  Transect=as.factor(Transect),
  Reef=as.factor(Reef),
  T_ID=as.factor(T_ID),
  Length_mm=as.double(Length_mm),
  Survey=as.factor(Survey),
  Site=as.factor(Site))

```


# Match with MERMAID database for LW constants
```{r}

mermaid.species <-read_csv("../0_data/MERMAID_data/fish-species_GFG.csv")

mermaid.species<-mermaid.species %>% mutate(
  Species_name = paste(Genus,Species, sep = " ")
)

fish<-fish %>% mutate(
  Species_name = paste(Genus,Species, sep = " ")
)

fish %>% anti_join(mermaid.species, by = "Species_name")

```


#Fix all the issues! Working off my updated mermaid list data base
```{r}

# All NA species get "sp"
fish$Species[is.na(fish$Species)]<- "sp"
# spp <- sp in Species
fish$Species[fish$Species == "spp"] <- "sp"

# Oxycheilinus diagrammus <- digramma. 
fish <- fish %>% 
  mutate(
    Species = case_when(Genus == "Oxycheilinus" & Species == "digrammus" ~ "digramma",
                        TRUE ~ as.character(Species))
  )

# Chaetodon mertensii - spelling corrected
# martensii - spelling corrected
# meretensii - spelling corrected
fish <- fish %>% 
  mutate(
    Species = case_when(Genus == "Chaetodon" & Species == "mertensi" ~ "mertensii",
                        TRUE ~ as.character(Species))
  )

fish <- fish %>% 
  mutate(
    Species = case_when(Genus == "Chaetodon" & Species == "martensi" ~ "mertensii",
                        TRUE ~ as.character(Species))
  )

fish <- fish %>% 
  mutate(
    Species = case_when(Genus == "Chaetodon" & Species == "meretensii" ~ "mertensii",
                        TRUE ~ as.character(Species))
  )

fish <- fish %>% 
  mutate(
    Species = case_when(Genus == "Chaetodon" & Species == "martensii" ~ "mertensii",
                        TRUE ~ as.character(Species))
  )

# Gymnocranius eunanus <- euanus - spelling
fish <- fish %>% 
  mutate(
    Species = case_when(Genus == "Gymnocranius" & Species == "eunanus" ~ "euanus",
                        TRUE ~ as.character(Species))
  )

# Anampses feminius - spelling of femininus
fish <- fish %>% 
  mutate(
    Species = case_when(Genus == "Anampses" & Species == "feminius" ~ "femininus",
                        TRUE ~ as.character(Species))
  )


# Acanthurus nubilis  - spelling <- nubilus
fish <- fish %>% 
  mutate(
    Species = case_when(Genus == "Acanthurus" & Species == "nubilis" ~ "nubilus",
                        TRUE ~ as.character(Species))
  )

# Diagramma pictum/pictum labiosum  - just pictum on MERMAID
fish <- fish %>% 
  mutate(
    Species = case_when(Genus == "Diagramma" & Species == "pictum labiosum" ~ "pictum",
                        TRUE ~ as.character(Species))
  )


# and a spelling mistake - starki
fish <- fish %>% 
  mutate(
    Species = case_when(Genus == "Chrysiptera" & Species == "starki" ~ "starcki",
                        TRUE ~ as.character(Species))
  )


# Sufflamen chrysopterus - spelling <- chrysopterum
fish <- fish %>% 
  mutate(
    Species = case_when(Genus == "Sufflamen" & Species == "chrysopterus" ~ "chrysopterum",
                        TRUE ~ as.character(Species))
  )

# Chaetodon pelwensis - spelling <-pelewensis
fish <- fish %>% 
  mutate(
    Species = case_when(Genus == "Chaetodon" & Species == "pelwensis" ~ "pelewensis",
                        TRUE ~ as.character(Species))
  )


# Chromis xanthurus - spelling <- xanthura
fish <- fish %>% 
  mutate(
    Species = case_when(Genus == "Chromis" & Species == "xanthurus" ~ "xanthura",
                        TRUE ~ as.character(Species))
  )

# Ctenochaetus nigricans - error. Video checked - should be acanthurus nigricans
fish <- fish %>% 
  mutate(
    Genus = case_when(Genus == "Ctenochaetus" & Species == "nigricans" ~ "Acanthurus",
                        TRUE ~ as.character(Genus))
  )

# Carcharhinus amblyrhynchoides should be amblyrhincos
fish <- fish %>% 
  mutate(
    Species = case_when(Genus == "Carcharhinus" & Species == "amblyrhynchoides" ~ "amblyrhynchos",
                        TRUE ~ as.character(Species))
  )

# Spelling issues with Genicanthus watanabei
fish %>% filter(
  str_detect(Species, pattern = "wata")
  )

fish <- fish %>% 
  mutate(
    Species = case_when(Genus == "Genicanthus" & str_detect(Species, "wata") ~ "watanabei",
                        TRUE ~ as.character(Species))
  )

# Pseudochromis flavipinnis < Chrysiptera flavipinnis
fish <- fish %>% 
  mutate(
    Genus = case_when(Genus == "Pseudochromis" & Species == "flavipinnis" ~ "Chrysiptera",
                        TRUE ~ as.character(Genus))
  )

# Cephalopholis leopardus < missing genus
fish <- fish %>% 
  mutate(
    Genus = case_when(Family == "Serranidae" & Species == "leopardus" ~ "Cephalopholis",
                        TRUE ~ as.character(Genus))
  )


# Chlorurus  spilurus should be used instead of C.sordidus
fish <- fish %>% 
  mutate(
    Species = case_when(Genus == "Chlorurus" & Species == "sordidus" ~ "spilurus",
                        TRUE ~ as.character(Species))
  )

```

# Pre merge second check

```{r}
# Refresh species_name after above changes
fish<-fish %>% mutate(
  Species_name = paste(Genus,Species, sep = " ")
)


fish %>% anti_join(mermaid.species, by = "Species_name")

```


# Same species match wrangle but using the rfishbase list

```{r}

fishbase.data <-read_csv("../0_data/fishbase_data/fishbase_list.csv")
fishbase.data<- fishbase.data %>% rename(
  Species_name = Species
)

fish %>% anti_join(fishbase.data, by = "Species_name")

```

#Fix issues
```{r}
# Pseudocaranx georgianus is not accepted name <- Pseudocaranx dentex
fish <- fish %>% 
  mutate(
    Species = case_when(Genus == "Pseudocaranx" & Species == "georgianus" ~ "dentex",
                        TRUE ~ as.character(Species))
  )
# Zebrasonma veliferum is not accepted name <- Zebrasoma velifer
fish <- fish %>% 
  mutate(
    Species = case_when(Genus == "Zebrasoma" & Species == "veliferum" ~ "velifer",
                        TRUE ~ as.character(Species))
  )
```

```{r}
# Refresh species_name after changes
fish<-fish %>% mutate(
  Species_name = paste(Genus,Species, sep = " ")
)


fish %>% anti_join(fishbase.data, by = "Species_name")

# All that remains should be observations only identified as "sp"

# Pseudanthias englehardorum is an issue - should be englehardi?
fish <- fish %>% 
  mutate(
    Species = case_when(Genus == "Pseudanthias" & Species == "englehardorum" ~ "engelhardi",
                        Genus == "Pseudanthias" & Species == "engelhardorum" ~ "engelhardi",
                        TRUE ~ as.character(Species))
  )

# Naso thynoides
fish <- fish %>% 
  mutate(
    Species = case_when(Genus == "Naso" & Species == "thynoides" ~ "thynnoides",
                        TRUE ~ as.character(Species))
  )

# Refresh species names
fish<-fish %>% mutate(
  Species_name = paste(Genus,Species, sep = " ")
)


fish %>% anti_join(fishbase.data, by = "Species_name")
```
# NA sp needs modifying
```{r}

fish <- fish %>% mutate(
  Species_name = case_when(Species_name == "NA sp" ~ paste(Family, Species, sep= " "),
                           TRUE ~ Species_name))

# Check again
fish %>% anti_join(fishbase.data, by = "Species_name")

# All that remains should be observations only identified as "sp"

```



# If duplicates have been generated check
```{r}

distinct(fish)
unique(fish)

```

# Merge with mermaid data
The fishbase data set was good for validating accepted species names, further spelling checks etc but had some issues with assigning the correct a and b constants. Firstly, there are multiple values for a given species - which is correct? They also seemed very different from the Mermaid ones and often the ones given on the fish's page on fishbase.org. For these reasons using the mermaid database.

KEEP YOUR MERMAID DATABASE UPTO DATE! If you need to add species, look up their a and b constants and enter manually to the csv. Keep this version up to date in all locations.

```{#r}
# MERMAID merge for a and b constants

merged.data <- fish %>% 
  left_join(mermaid.species, by ="Species_name")

duplicated(mermaid.species$Species_name)

mermaid.species[which(duplicated(mermaid.species[,c('Species_name')])==TRUE),]

```

```{r}

# Work going forward will now always use my own custom fish list. This has been updated with many Coral Sea Species not originally on the list and includes biomass constants and max length. 

mermaid.species<- mermaid.species %>% select(Species_name,`Biomass Constant A`,`Biomass Constant B`, `Max Length (cm)`, `Trophic Group`, `Trophic Level`)

merged.data <- fish %>% 
  left_join(mermaid.species, by ="Species_name")

# See which species are still missing a and b values. We expect anything identified to just family or genus to have NA. This will be dealt with later on.

add.data<-merged.data %>% filter(is.na(`Biomass Constant A`)) %>% distinct(Species_name)

add.data$Species_name %>% str_subset(pattern = "sp$", negate = TRUE)

# \s is regex for white space - could also us that



# Check missing constants
merged.data %>% filter(is.na(`Biomass Constant A`))

# Check again, ensuring any missing values are just observations of family or genus level
add.data<-merged.data %>% filter(is.na(`Biomass Constant A`)) %>% distinct(Species_name)

add.data$Species_name %>% str_subset(pattern = "sp$", negate = TRUE)



```


# Biomass constants for obervations with just Family or genus
MERMAID data
```{r}

mermaid.genus <-read_csv("../0_data/MERMAID_data/fish-genera.csv")
mermaid.genus<- mermaid.genus %>% select(Family,Genus,`Biomass Constant A`,`Biomass Constant B`)

mermaid.family<- read_csv("../0_data/MERMAID_data/fish-families.csv")
mermaid.family<- mermaid.family %>% select(Family,`Biomass Constant A`,`Biomass Constant B`)


```

# Missing species
```{r}
#Check missing species
# Inspect the species variable and identify na or "sp" values 
merged.data %>% filter(
  is.na(Species) | Species == "sp" | Species == "sp2")

# Extract these observations as a separate df
sp.data<-merged.data %>% filter(
  is.na(Species)| Species == "sp"| Species == "sp2")

# Remove from original merged.data df - rebind later once corrected
#fish.1<-merged.data %>% filter(Species != "sp") 
merged.data <- merged.data %>% filter(Species != "sp") 
merged.data <- merged.data %>% filter(Species != "sp2") 
```

# Mising genera

```{r}
#Check missing genera from the sp.data
# Inspect the species variable and identify na 

sp.data %>% filter(is.na(Genus))
summary(is.na(sp.data$Genus))

# Extract these observations as a separate df
genus.data<-sp.data %>% filter(
  is.na(Genus))

# Remove from original merged.data df - rebind later once corrected
sp.data <- sp.data %>% filter(!is.na(Genus))

```

# Merge with MERMAID attribute data for mising species
```{r}

merged.sp <- sp.data %>% 
  left_join(mermaid.genus, by =c("Family","Genus")) %>% 
              mutate(`Biomass Constant A` = coalesce (`Biomass Constant A.x`, `Biomass Constant A.y`),
                     `Biomass Constant B` = coalesce (`Biomass Constant B.x`, `Biomass Constant B.y`)) %>% 
  select(-c(21,22,26,27))

```

# Error checks for na values
```{r}
merged.sp %>% filter(is.na(`Biomass Constant A`)) # Caesio
merged.sp %>% anti_join(mermaid.genus, by = c("Family")) #Caesioinidae
                                                         #Caesionidae - incorrect spelling

sp.data <- sp.data %>% 
  mutate(
    Family = case_when(Family == "Caesionidae" ~ "Caesioinidae",
                        TRUE ~ as.character(Family))
  )

# re-run 
merged.sp <- sp.data %>% 
  left_join(mermaid.genus, by =c("Family","Genus")) %>% 
              mutate(`Biomass Constant A` = coalesce (`Biomass Constant A.x`, `Biomass Constant A.y`),
                     `Biomass Constant B` = coalesce (`Biomass Constant B.x`, `Biomass Constant B.y`)) %>% 
 select(-c(22,23,27,28))
```

# Merge with MERMAID attribute data for misiing genera
```{r}
merged.genera <- genus.data %>% 
  left_join(mermaid.family, by = "Family") %>% 
              mutate(`Biomass Constant A` = coalesce (`Biomass Constant A.x`, `Biomass Constant A.y`),
                     `Biomass Constant B` = coalesce (`Biomass Constant B.x`, `Biomass Constant B.y`)) %>% 
 select(-c(22,23,27,28))
```

# Row bind all data back together

```{r}
fish.1<- rbind(merged.data, merged.genera, merged.sp)

# Final check that every observation has a and or b constants - ALL should be false
summary(is.na(fish.1$`Biomass Constant A`))
fish.1 %>% filter(is.na(`Biomass Constant A`))


# Check again, ensuring any missing values are just observations of family or genus level
missing.data<-fish.1 %>% filter(is.na(`Biomass Constant A`)) %>% distinct(Species_name)
missing.data$Species_name %>% str_subset(pattern = "sp$", negate = TRUE)
```

```{r}
# Write to csv

write.csv(fish.1, file= "../2_outputs/R0V_Fish_ab.csv", row.names = FALSE)

```


# Coalesce join function - much more elegant solution!
Still needs work
```{r}
coalesce_join <- function(x, y, 
                          by = NULL, suffix = c(".x", ".y"), 
                          join = dplyr::full_join, ...) {
    joined <- join(x, y, by = by, suffix = suffix, ...)
    # names of desired output
    cols <- union(names(x), names(y))
    
    to_coalesce <- names(joined)[!names(joined) %in% cols]
    suffix_used <- suffix[ifelse(endsWith(to_coalesce, suffix[1]), 1, 2)]
    # remove suffixes and deduplicate
    to_coalesce <- unique(substr(
        to_coalesce, 
        1, 
        nchar(to_coalesce) - nchar(suffix_used)
    ))
    
    coalesced <- purrr::map_dfc(to_coalesce, ~dplyr::coalesce(
        joined[[paste0(.x, suffix[1])]], 
        joined[[paste0(.x, suffix[2])]]
    ))
    names(coalesced) <- to_coalesce
    
    dplyr::bind_cols(joined, coalesced)[cols]
}

```


# Example Merge using new function for missing genera
```{r}

merged.genera<-coalesce_join(genus.data, mermaid.family, by = 'Family')

# Seems to add all the extra rows from df2 that aren't in df1? Needs tweaking.
# Change initial function call to full_join to left join? I think

```
