==> 1_scripts/04_fish_average_lengths.Rmd <==
---
title: "fish_average_lengths"
author: "Gemma Galbraith"
date: "28/05/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

The purpose of this script is to look for fish observations that do not have a length estimate and apply the average length for that species from all other observations.

#HouseKeeping
```{R}
rm(list=ls())
#dev.off()
getwd()

```

#load libraries
```{r}
library(tidyverse)
library(rfishbase)
```

#Import fish data
```{r}
fish <-read_csv("../2_outputs/R0V_Fish_ab.csv")

fish <- fish %>% mutate(
  Transect=as.factor(Transect),
  Reef=as.factor(Reef),
  T_ID=as.factor(T_ID),
  Length_mm=as.double(Length_mm),
  Survey=as.factor(Survey),
  Site=as.factor(Site))

```

# Remove dives where no stereo was taken
- Herald's surprise
- Juliette Cay
- Edna Cay

```{r}

# Check correct sites are being pulled
fish %>% filter(Video == "Single") %>% distinct(Site)

# Extract these observations as a separate df
singles<-fish %>% filter(Video == "Single")

# Remove from original fish - rebind later once mean lengths have been calculated
fish <- fish %>% filter(Video == "Stereo")

```


# DF of average length for all species with length estimates
```{r}

means<-fish %>% group_by(Species_name) %>% 
  summarise(
    Mean_length = mean(na.omit(Length_mm))
  )

fish.2<-left_join(fish, means, by = "Species_name")

fish.2<-fish.2 %>%  mutate(
  Length_mm = coalesce(Length_mm,Mean_length))

# Check for missing length values now

no.values<-fish.2 %>% filter(is.na(Length_mm))  
fish.2 %>% filter(is.na(Mean_length)) # some species had no other observations so a mean cannot be calculated

#remove from fish.2
fish.2<- fish.2 %>% filter(
  !is.na(Length_mm)
)

```

# Dealing with remaining na values - check which observations/species have no length measurements

```{r}

unique(no.values$Species_name)

```

# Errors spotted

```{r}

# Need to reapply those means
fish.2<-left_join(fish, means, by = "Species_name")

fish.2<-fish.2 %>%  mutate(
  Length_mm = coalesce(Length_mm,Mean_length))


# Check for missing length values now

no.values<-fish.2 %>% filter(is.na(Length_mm))  
fish.2<-fish.2 %>% filter(!is.na(Length_mm))


```

# Apply a mean length 
Could calculate these based on the in-situ data? Otherwise apply an estimate?

- Option 1 - Manual Entry
```{r #}
unique(no.values$Species_name)

null.species<-unique(no.values$Species_name)
null.species

# est.mean.lengths applied as "common length" from fishbase.se



```

- Option 2 - fishbase species table
```{r}
fishbase<-rfishbase::species()
write.csv(fishbase, file= "../0_data/fishbase_data/fishbase_species.csv", row.names = FALSE)

fishbase<-fishbase %>% select(Species, CommonLength) %>% 
  rename(Species_name = Species)

```

# Match no.values with fishbase species
```{r}
no.values<- no.values %>% 
  left_join(fishbase, by= "Species_name")
  

no.values<-no.values %>% 
    mutate(Length_mm = coalesce(Length_mm,CommonLength))

no.values<-no.values %>% select(-CommonLength)

```

# row bind species with na values back to main fish.2 data

```{r}
fish<-rbind(fish.2, no.values) %>% 
  select(-Mean_length)


# Check everything has required length and a/b constants for biomass calculations
fish %>% filter(is.na(`Biomass Constant A`))
fish %>% filter(is.na(`Biomass Constant B`))
fish %>% filter(is.na(Length_mm))

# fishbase list is still missing some values for common lengths. Missing species added manually.
no.values$Length_mm[no.values$Species_name == "Rhabdamia gracilis"] <- 50
no.values$Length_mm[no.values$Species_name == "Aluterus scriptus"] <- 500
no.values$Length_mm[no.values$Species_name == "Halichoeres chrysus"] <- 100
no.values$Length_mm[no.values$Species_name == "Siganus argenteus"] <- 250
no.values$Length_mm[no.values$Species_name == "Thalassoma hardwicke"] <- 150
no.values$Length_mm[no.values$Species_name == "Pseudanthias pictilis"] <- 110
no.values$Length_mm[no.values$Species_name == "Rhinecanthus aculeatus"] <- 150
no.values$Length_mm[no.values$Species_name == "Chromis chrysura"] <- 100
no.values$Length_mm[no.values$Species_name == "Lutjanus adetii"] <- 300
no.values$Length_mm[no.values$Species_name == "Balistapus undulatus"] <- 260
no.values$Length_mm[no.values$Species_name == "Diagramma pictum"] <-  550
no.values$Length_mm[no.values$Species_name == "Acanthurus nigricauda"] <- 270
no.values$Length_mm[no.values$Species_name == "Stegastes fasciolatus"] <- 110
no.values$Length_mm[no.values$Species_name == "Cirrhilabrus scottorum"] <- 120  
no.values$Length_mm[no.values$Species_name == "Cirrhilabrus punctatus"] <- 120  
no.values$Length_mm[no.values$Species_name == "Cephalopholis miniata"] <-  300
no.values$Length_mm[no.values$Species_name == "Amphiprion chrysopterus"] <-  130
no.values$Length_mm[no.values$Species_name == "Chilomycterus reticulatus"] <-  300
no.values$Length_mm[no.values$Species_name == "Centropyge tibicen"] <-   150
no.values$Length_mm[no.values$Species_name == "Paracirrhites arcatus"] <-  150
no.values$Length_mm[no.values$Species_name == "Coris pictoides"] <-  130
no.values$Length_mm[no.values$Species_name == "Naso hexacanthus"] <- 500 
no.values$Length_mm[no.values$Species_name == "Arothron stellatus"] <-  540
no.values$Length_mm[no.values$Species_name == "Naso tonganus"] <-  500
no.values$Length_mm[no.values$Species_name == "Meiacanthus atrodorsalis"] <-  100
no.values$Length_mm[no.values$Species_name == "Chrysiptera tricincta"] <-  55
no.values$Length_mm[no.values$Species_name == "Halichoeres hortulanus"] <- 200
no.values$Length_mm[no.values$Species_name == "Aulostomus chinensis"] <-  600
no.values$Length_mm[no.values$Species_name == "Gymnosarda unicolor"] <- 1900 
no.values$Length_mm[no.values$Species_name == "Epibulus insidiator"] <-  450
no.values$Length_mm[no.values$Species_name == "Pseudocaranx dentex"] <- 400
no.values$Length_mm[no.values$Species_name == "Gymnocranius superciliosus"] <- 400
no.values$Length_mm[no.values$Species_name == "Rhinecanthus lunula"] <- 220
no.values$Length_mm[no.values$Species_name == "Dipterygonotus balteatus"] <- 120
no.values$Length_mm[no.values$Species_name == "Synodus sp"] <- 200
no.values$Length_mm[no.values$Species_name == "Chrysiptera taupou"] <- 70
no.values$Length_mm[no.values$Species_name == "Labroides sp"] <- 100

# Check again
no.values %>% filter(is.na(Length_mm))

# remerge
fish<-rbind(fish.2, no.values) %>% 
  select(-Mean_length)

fish %>% filter(is.na(Length_mm))

```

# gross length checks on species - visual
Inspect the larger values - validate e.g. a 1 meter long S.bursa!!!
```{r}

fish <- fish %>% 
  mutate(
    Length_mm = case_when(Length_mm == 1763.2370 & T_ID == "Feb_21_Willis_N1_T9" ~ 176,
                        TRUE ~ as.double(Length_mm))
  )

```

# Bind the single video surveys back in
```{r}
fish<-rbind(fish, singles)


```

# Write csv

```{r}

write.csv(fish, file= "../2_outputs/ROV_Fish_lengths.csv", row.names = FALSE)

```
