==> 1_scripts/05_ROV_Trophic_group_assignment.Rmd <==
---
title: "10_Trophic_group_assignment"
author: "GF Galbraith"
date: '2022-08-17'
output: html_document
---
---
title: "Assigning Missing Trophic Attributes"
output: html_document
date: '2022-07-04'
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Using the cleaned BRUV "BRUV_Fish_lengths.csv" data generated by script 04, this markdown assigns missing trophic attributes to species/observations. The resulting data frame is the final BRUV fish data used in most other analyses and figure plotting.

#HouseKeeping
```{R}
rm(list=ls())
#dev.off()
getwd()

```

#load libraries
```{r}
library(tidyverse)
library(rfishbase)
```

# Data import
```{r}
fish <-read_csv("../2_outputs/ROV_Fish_lengths.csv")

fish<- fish %>% mutate(
 `Trophic Group` = as.factor(`Trophic Group`),
#  `Functional Group` = as.factor(`Functional Group`),
  Species_name = as.factor(Species_name)
)

```

# Look for NA vlaues in `Trophic Group`.
```{r}

missing<-fish%>%
  filter(is.na(`Trophic Group`))%>%
  distinct(Species_name, .keep_all = TRUE)

missing$Species_name
```

# Look at the levels available assigned from Mermaid
```{r}
levels(fish$`Trophic Group`)
levels(fish$`Functional Group`)


```

# Tidy assign
Potentially not possible :(

# Mannual assign - ADD TO MERMAID master csv
```{r}

# Gnathodentex aureolineatus
fish$`Trophic Group`[fish$Species_name == "Gnathodentex aureolineatus"] <- "invertivore-mobile"

# Carangidae sp
fish$`Trophic Group`[fish$Species_name == "Carangidae sp"] <- "piscivore"

#Labridae sp      
fish$`Trophic Group`[fish$Species_name == "Labridae sp"] <- "invertivore-mobile"

#Pomacentridae sp  
fish$`Trophic Group`[fish$Species_name == "Pomacentridae sp"] <- "planktivore"

#Clupeidae sp             
fish$`Trophic Group`[fish$Species_name == "Clupeidae sp"] <- "planktivore"

#Myripristis sp             
fish$`Trophic Group`[fish$Species_name == "Myripristis sp"] <- "planktivore"

#Meiacanthus sp             
fish$`Trophic Group`[fish$Species_name == "Meiacanthus sp"] <- "planktivore"

#Scarus sp                 
fish$`Trophic Group`[fish$Species_name == "Scarus sp"] <- "herbivore-detritivore"

#Chromis sp                 
fish$`Trophic Group`[fish$Species_name == "Chromis sp"] <- "planktivore"

#Pseudanthias sp            
fish$`Trophic Group`[fish$Species_name == "Pseudanthias sp"] <- "planktivore"

#Thalassoma sp              
fish$`Trophic Group`[fish$Species_name == "Thalassoma sp"] <- "invertivore-mobile"

#Anampses sp               
fish$`Trophic Group`[fish$Species_name == "Anampses sp"] <- "invertivore-mobile"

#Sargocentron sp            
fish$`Trophic Group`[fish$Species_name == "Sargocentron sp"] <- "piscivore"

#Cephalopholis sp           
fish$`Trophic Group`[fish$Species_name == "Cephalopholis sp"] <- "piscivore"

#Halichoeres sp             
fish$`Trophic Group`[fish$Species_name == "Halichoeres sp"] <- "invertivore-mobile"

#Myliobatis sp             
fish$`Trophic Group`[fish$Species_name == "Myliobatis sp"] <- "invertivore-sessile"

#Pomacentrus sp             
fish$`Trophic Group`[fish$Species_name == "Pomacentrus sp"] <- "planktivore"

#Acanthurus sp   
fish$`Trophic Group`[fish$Species_name == "Acanthurus sp"] <- "herbivore-detritivore"

#Ctenochaetus sp            
fish$`Trophic Group`[fish$Species_name == "Ctenochaetus sp"] <- "herbivore-detritivore"

#Centropyge sp             
fish$`Trophic Group`[fish$Species_name == "Centropyge sp"] <- "herbivore-detritivore"

#Lethrinus sp               
fish$`Trophic Group`[fish$Species_name == "Lethrinus sp"] <- "piscivore"

#Pentapodus sp              
fish$`Trophic Group`[fish$Species_name == "Pentapodus sp"] <- "invertivore-mobile"

#Stegastes sp               
fish$`Trophic Group`[fish$Species_name == "Stegastes sp"] <- "herbivore-detritivore"

#Cirrhilabrus sp           
fish$`Trophic Group`[fish$Species_name == "Cirrhilabrus sp"] <- "planktivore"

#Hoplolatilus sp            
fish$`Trophic Group`[fish$Species_name == "Hoplolatilus sp"] <- "planktivore"

#Synodus sp                 
fish$`Trophic Group`[fish$Species_name == "Synodus sp"] <- "piscivore"

#Coris sp                   
fish$`Trophic Group`[fish$Species_name == "Coris sp"] <- "invertivore-mobile"

#Pseudocheilinus sp        
fish$`Trophic Group`[fish$Species_name == "Pseudocheilinus sp"] <- "invertivore-mobile"

#Canthigaster sp            
fish$`Trophic Group`[fish$Species_name == "#Canthigaster sp"] <- "invertivore-mobile"

#Pseudodax sp               
fish$`Trophic Group`[fish$Species_name == "Pseudodax sp"] <- "invertivore-mobile"

#Ptereleotris sp            
fish$`Trophic Group`[fish$Species_name == "Ptereleotris sp"] <- "planktivore"

#Parupeneus sp             
fish$`Trophic Group`[fish$Species_name == "Parupeneus sp"] <- "invertivore-mobile"

#Rhabdamia sp               
fish$`Trophic Group`[fish$Species_name == "Rhabdamia sp"] <- "planktivore"

#Caesio sp                  
fish$`Trophic Group`[fish$Species_name == "Caesio sp"] <- "planktivore"

#Pseudochromis sp           
fish$`Trophic Group`[fish$Species_name == "Pseudochromis sp"] <- "planktivore"

#Pseudocoris sp            
fish$`Trophic Group`[fish$Species_name == "Pseudocoris sp"] <- "planktivore"

#Chaetodontoplus sp         
fish$`Trophic Group`[fish$Species_name == "Chaetodontoplus sp"] <- "invertivore-mobile"

#Sufflamen sp               
fish$`Trophic Group`[fish$Species_name == "Sufflamen sp"] <- "invertivore-mobile"

#Chaetodon sp               
fish$`Trophic Group`[fish$Species_name == "Chaetodon sp"] <- "herbivore-detritivore"

#Genicanthus sp            
fish$`Trophic Group`[fish$Species_name == "Genicanthus sp"] <- "planktivore"

#Chromis chrysura           
fish$`Trophic Group`[fish$Species_name == "Chromis chrysura"] <- "planktivore"

#Labroides sp               
fish$`Trophic Group`[fish$Species_name == "Labroides sp"] <- "invertivore-mobile"

#Pseudanthias sp2           
fish$`Trophic Group`[fish$Species_name == "Pseudanthias sp2"] <- "planktivore"

#Choerodon sp              
fish$`Trophic Group`[fish$Species_name == "Choerodon sp"] <- "invertivore-mobile"

#Aulostomus sp              
fish$`Trophic Group`[fish$Species_name == "Aulostomus sp"] <- "piscivore"

#Rhinecanthus sp           
fish$`Trophic Group`[fish$Species_name == "Rhinecanthus sp"] <- "invertivore-mobile"

#Pomacentrus sp2        
fish$`Trophic Group`[fish$Species_name == "Pomacentrus sp2"] <- "planktivore"

#Canthigaster sp        
fish$`Trophic Group`[fish$Species_name == "Canthigaster sp"] <- "invertivore-mobile"

```

# Check for NAs again
```{r}
fish%>%
  filter(is.na(`Trophic Group`))%>%
  distinct(Species_name, .keep_all = TRUE)
```

# Apply monitoring functional groups
```{r}
monitoring <-read_csv("../0_data/Monitoring_species.csv")
monitoring<- monitoring %>% select(Family, Species,`Functional group`)%>%
  mutate(Species_name = as.factor(Species),
         `Functional group` = as.factor(`Functional group`))

monitoring<- monitoring %>%select(Family, `Species_name`, `Functional group`)

```

```{r}
fish2<-left_join(fish, monitoring, by = c("Species_name"))

```

# Find Na values

```{r}

fish2 %>%
  filter(is.na(`Functional group`))%>%
  distinct(Species_name, .keep_all = TRUE)%>%
  pull(Species_name)

# List extracted - fix in monitoring csv

missing<-fish2 %>%
  filter(is.na(`Functional group`))%>%
  distinct(Species_name, .keep_all = TRUE)%>%
  mutate( Family.x = as.factor(Family.x)) %>% 
  select(Family.x, Species_name)
  
  
  #pull(Species_name) %>% as.data.frame()


write.csv(missing, file= "../2_outputs/missing_funcational_groups.csv", row.names = FALSE)

levels(monitoring$`Functional group`)
levels(missing$Family.x)

```

# Add correct Functional groups as per monitoring database
This has been done in the monitoring csv file, now updated and called monitoring_species_GFG

```{r}

monitoring2 <-read_csv("../0_data/Monitoring_species_GFG.csv")
monitoring2<- monitoring2 %>% select(Family, Species,`Functional group`)%>%
  mutate(Species_name = as.factor(Species),
         `Functional group` = as.factor(`Functional group`))

monitoring2<- monitoring2 %>%select(Family, `Species_name`, `Functional group`)

# Seems to have duplicate values


```
# Check NA values again
```{r}
fish3<-left_join(fish, monitoring2, by = c("Species_name"))

# This merge adds observations? Should have 3703 not 3975. monitoring2 must have duplicated species_names
duplicated(monitoring2)
monitoring3<-distinct(monitoring2,Species_name,.keep_all = TRUE)

fish3<-left_join(fish, monitoring3, by = c("Species_name"))

# Check for NA in functional group again
fish3 %>%
  filter(is.na(`Functional group`))%>%
  distinct(Species_name, .keep_all = TRUE)%>%
  pull(Species_name)

```

# FAMILY DISCREPANCIES
When merging the monitoring species list for functional groups, joining by family would generate NA values. This makes me think that we have discrepancies between the family designations. Maybe Family vs subfamily? E.g. Epinephelus sp, Serranidae vs Epinephelidae?


# Check for NA values in Family

# Find Na values

```{r}

families<-fish3 %>%
  select(Species_name,Family.x, Family.y)


# Where do family names NOT match

no.match <- families %>% subset(Family.x != Family.y) %>%
  distinct(Species_name, .keep_all = TRUE)
  
```
# Manual changes made in csv

Ostracidae spelt wrong in monitoring list --> Ostraciidae
Dasyatiidae spelt wrong in monitoring list --> Dasyatidae
Congidae spelt wrong in monitoring list --> Congridae
Malacantidae spelt wrong in monitoring list --> Malacanthidae


# Chnages made in R
Gymnocranius euanus - has one obs with wrong family
```{r}
fish3$Family.x[fish3$Species_name == "Gymnocranius euanus"] <- "Lethrinidae"

```

The other disagreements are Family/subfamily divisions
Scaridae / Labridae (Scarinae)
Epinephelidae/ Serranidae

Chosen to use Family designation from now corrected monitoring_GFG species list (family.y)

```{r}

fish3<-fish3%>%rename (Family = Family.y)

fish3<-fish3%>%select( -Family.x)

fish3<-fish3%>%rename (`Functional Group` = `Functional group`)

```

# Write to CSV as Final BRUV Fish data
```{r}
write.csv(fish3, file= "../2_outputs/ROV_Fish_Final.csv", row.names = FALSE)
```
