==> 1_scripts/07_diversity_calculations.Rmd <==
---
title: "06_diversity_calculations"
author: "Gemma Galbraith"
date: "30/05/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#HouseKeeping
```{R}
rm(list=ls())
#dev.off()
getwd()

```

#load libraries
```{r}
library(tidyverse)
library(rfishbase)
library(vegan)
```

# Data import
```{r}
fish <-read_csv("../2_outputs/ROV_Fish_Final.csv")

fish <- fish %>% mutate(
  Transect=as.factor(Transect),
  Reef=as.factor(Reef),
  T_ID=as.factor(T_ID),
  Length_mm=as.double(Length_mm),
  Survey=as.factor(Survey),
  Site=as.factor(Site),
  Comment = as.character(Comment),
  Date = as.Date(Date),
  `Trophic Group` = as.factor(`Trophic Group`),
  `Functional Group` = as.factor(`Functional Group`),
  Habitat = as.factor(Habitat),
  Aspect = as.factor(Aspect))
```

# Create a species matrix in wide format for vegan functions and diversity indices

```{r long to wide}
# Select only the unique ID, .species and count
fish.mat<-fish %>% select(T_ID, Species_name, Number)

#fish.mat<-fish.mat %>% relocate(Number, .after = Species_name)

fish.mat<-fish.mat %>% group_by(T_ID,Species_name) %>% 
  summarise(Number=sum(Number)) %>% ungroup()

fish.mat<-fish.mat %>% pivot_wider(names_from = Species_name, values_from = Number) %>% 
  replace(is.na(.),0)
```

```{r Diversity indices}

# Extract a species matrix - just species name and count
spec.mat<-fish.mat[,-1]

# use the species matrix to calculate the indices
fish.mat$richness <- rowSums(spec.mat>0)
fish.mat$abundance<- apply(spec.mat,1, sum)
fish.mat$shannon <- diversity(spec.mat) # shannon is default
fish.mat$simpson<-diversity(spec.mat, index = "simpson")

```

# Select diversity and TID
```{r}
diversity.data<-fish.mat %>% select(T_ID, richness,abundance, shannon,simpson)

transect.data<-fish %>% select(Survey,Reef,Site,Transect,T_ID,Date, Lat, Long, Depth, Temperature, Habitat, Aspect, Benthic_Complexity)

transect.data<-distinct(transect.data, T_ID, .keep_all = TRUE)

diversity.data<-left_join(diversity.data, transect.data, by="T_ID")
```

# Write CSV
```{r}

write.csv(diversity.data, file= "../2_outputs/ROV_fish_diversity.csv", row.names = FALSE)

```
