==> 1_scripts/10_BRUV_trophic_attributes.Rmd <==
---
title: "Assigning Missing Trophic Attributes"
output: html_document
date: '2022-07-04'
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Using the cleaned BRUV "BRUV_Fish_ab.csv" data generated by script 02, this markdown assigns missing trophic attributes to species/observations. The resulting data frame is the final BRUV fish data used in most other analyses and figure plotting.

#HouseKeeping
```{R}
rm(list=ls())
#dev.off()
getwd()

```

#load libraries
```{r}
library(tidyverse)
library(rfishbase)
```

# Data import
```{r}
fish <-read_csv("../2_outputs/BRUV_Fish_ab.csv")

fish<- fish %>% mutate(
  `Trophic Group` = as.factor(`Trophic Group`),
  `Functional Group` = as.factor(`Functional Group`),
  Species_name = as.factor(Species_name)
)

```

# Look for NA vlaues in `Trophic Group`.
```{r}

fish%>%
  filter(is.na(`Trophic Group`))%>%
  distinct(Species_name, .keep_all = TRUE)
```

# Look at the levels available assigned from Mermaid
```{r}
levels(fish$`Trophic Group`)
levels(fish$`Functional Group`)
```

# Mannual assign - ADD TO MERMAID master csv
```{r}
# Epinephelus polyphekadion 
fish$`Trophic Group`[fish$Species_name == "Epinephelus polyphekadion"] <- "piscivore"
fish$`Functional Group`[fish$Species_name == "Epinephelus polyphekadion"] <- "pisci-invertivore"

# Pervagor alternans
fish$`Trophic Group`[fish$Species_name == "Pervagor alternans"] <- "omnivore"
fish$`Functional Group`[fish$Species_name == "Pervagor alternans"] <- "micro-invertivore"

# Pterocaesio marri
fish$`Trophic Group`[fish$Species_name == "Pterocaesio marri"] <- "planktivore"
fish$`Functional Group`[fish$Species_name == "Pterocaesio marri"] <- "planktivore"

# Aluterus scriptus
fish$`Trophic Group`[fish$Species_name == "Aluterus scriptus"] <- "omnivore"
fish$`Functional Group`[fish$Species_name == "Aluterus scriptus"] <- "micro-invertivore"

# Naso sp
fish$`Trophic Group`[fish$Species_name == "Naso sp"] <- "planktivore"
fish$`Functional Group`[fish$Species_name == "Naso sp"] <- "planktivore"

# Epinephelus maculatus
fish$`Trophic Group`[fish$Species_name == "Epinephelus maculatus"] <- "piscivore"
fish$`Functional Group`[fish$Species_name == "Epinephelus maculatus"] <- "pisci-invertivore"

# Epinephelus fasciatus
fish$`Trophic Group`[fish$Species_name == "Epinephelus fasciatus"] <- "invertivore-mobile"
fish$`Functional Group`[fish$Species_name == "Epinephelus fasciatus"] <- "macro-invertivore"

# Chromis chrysura
fish$`Trophic Group`[fish$Species_name == "Chromis chrysura"] <- "planktivore"
fish$`Functional Group`[fish$Species_name == "Chromis chrysura"] <- "planktivore"

# Epinephelus tauvina
fish$`Trophic Group`[fish$Species_name == "Epinephelus tauvina"] <- "piscivore"
fish$`Functional Group`[fish$Species_name == "Epinephelus tauvina"] <- "pisci-invertivore"

# Gnathodentex aureolineatus
fish$`Trophic Group`[fish$Species_name == "Gnathodentex aureolineatus"] <- "invertivore-mobile"
fish$`Functional Group`[fish$Species_name == "Gnathodentex aureolineatus"] <- "macro-invertivore"

# Pterocaesio digramma
fish$`Trophic Group`[fish$Species_name == "Pterocaesio digramma"] <- "planktivore"
fish$`Functional Group`[fish$Species_name == "Pterocaesio digramma"] <- "planktivore"

# Epinephelus cyanopodus
fish$`Trophic Group`[fish$Species_name == "Epinephelus cyanopodus"] <- "piscivore"
fish$`Functional Group`[fish$Species_name == "Epinephelus cyanopodus"] <- "pisci-invertivore"

# Pseudalutarius nasicornis
fish$`Trophic Group`[fish$Species_name == "Pseudalutarius nasicornis"] <- "invertivore-sessile"
fish$`Functional Group`[fish$Species_name == "Pseudalutarius nasicornis"] <- "spongivore"

# Cirrhilabrus laboutei
fish$`Trophic Group`[fish$Species_name == "Cirrhilabrus laboutei"] <- "planktivore"
fish$`Functional Group`[fish$Species_name == "Cirrhilabrus laboutei"] <- "planktivore"
fish$Family[fish$Species_name == "Cirrhilabrus laboutei"] <- "Labridae"

# Gymnocranius euanus
fish$`Trophic Group`[fish$Species_name == "Gymnocranius euanus"] <- "invertivore-mobile"
fish$`Functional Group`[fish$Species_name == "Gymnocranius euanus"] <- "macro-invertivore"

```

# Check for NAs again
cccc

# Apply monitoring functional groups
```{r}
monitoring <-read_csv("../0_data/Monitoring_species.csv")
monitoring<- monitoring %>% select(Family, Species,`Functional group`)%>%
  mutate(Species_name = as.factor(Species),
         `Functional Group.X` = as.factor(`Functional group`))

monitoring<- monitoring %>%select(Family, `Species_name`, `Functional Group.X`)
```

```{r}
fish2<-left_join(fish, monitoring, by = c("Species_name"))

```

# Find Na values

```{r}

fish2 %>%
  filter(is.na(`Functional Group.X`))%>%
  distinct(Species_name, .keep_all = TRUE)%>%
  pull(Species_name)

# List extracted

```

# Add correct Functional groups as per monitoring database
This has been done in the monitoring csv file, now updated and called monitoring_species_GFG

```{r}

monitoring2 <-read_csv("../0_data/Monitoring_species_GFG.csv")
monitoring2<- monitoring2 %>% select(Family, Species,`Functional group`)%>%
  mutate(Species_name = as.factor(Species),
         `Functional Group.X` = as.factor(`Functional group`))

monitoring2<- monitoring2 %>%select(Family, `Species_name`, `Functional Group.X`)
```
# Check NA values again
```{r}
fish3<-left_join(fish, monitoring2, by = c("Species_name"))

fish3 %>%
  filter(is.na(`Functional Group.X`))%>%
  distinct(Species_name, .keep_all = TRUE)%>%
  pull(Species_name)

```

# FAMILY DISCREPANCIES
When merging the monitoring species list for functional groups, joinging by family would generate NA values. This makes me think that we have discrepancies between the family designations. Maybe Family vs subfamily? E.g. Epinephelus sp, Serranidae vs Epinephelidae?


# Check for NA values in Family
# Find Na values

```{r}

families<-fish3 %>%
  select(Species_name,Family.x, Family.y)


# Where do family names NOT match

no.match <- families %>% subset(Family.x != Family.y) %>%
  distinct(Species_name, .keep_all = TRUE)
  
```
# Manual changes made in csv

Ostracidae spelt wrong in monitoring list --> Ostraciidae
Dasyatiidae spelt wrong in monitoring list --> Dasyatidae
Congidae spelt wrong in monitoring list --> Congridae
Malacantidae spelt wrong in monitoring list --> Malacanthidae


# Chnages made in R
Gymnocranius euanus - has one obs with wrong family
```{r}
fish3$Family.x[fish3$Species_name == "Gymnocranius euanus"] <- "Lethrinidae"

```

The other disagreements are Family/subfamily divisions
Scaridae / Labridae (Scarinae)
Epinephelidae/ Serranidae

Chosen to use Family designation from now corrected monitoring_GFG species list

```{r}

fish3<-fish3%>%select(-Family, -Family.x)

fish3<-fish3%>%rename(Family = Family.y)

```

# Write to CSV as Final BRUV Fish data
```{r}
write.csv(fish3, file= "../2_outputs/BRUV_Fish_Final.csv", row.names = FALSE)

```
