==> 1_scripts/12_newspecies.Rmd <==
---
title: "12_newspecies"
author: "GF Galbraith"
date: '2022-06-24'
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

This script takes the fish observed on ROV or BRUV surveys and matches/anti-matches with the species list from the shallow monitoring surveys. ---> new species observed in deep surveys

Also matches with rfishbase list for depth ranges to give a list of new depth records from the study based on fishbase information.

#Housekeeping
```{}
rm(list=ls())
#dev.off()
getwd()
```

#Libraries
```{r}
library(tidyverse)
library(readr)
```

# Data import

```{r}
rov.fish <-read_csv("../2_outputs/ROV_Fish_Final.csv")  # ROV fish data including site info
bruv.fish <-read_csv("../2_outputs/BRUV_Fish_Final.csv")  # ROV fish data including site info

rov.fish <- rov.fish %>%select(T_ID,Survey, Reef, Site, Transect, Depth, Family, Genus, Species, Species_name)%>%mutate(
  Reef=as.factor(Reef),
  Survey=as.factor(Survey),
  Transect= as.factor(Transect),
  Site=as.factor(Site),
  T_ID_ = as.factor(T_ID),
  Depth = as.factor(Depth),
  Species_name = as.factor(Species_name))

bruv.fish<- bruv.fish %>% select(T_ID,Survey, Reef, Site, Drop, Depth, Family, Genus, Species, Species_name)%>% mutate(
  Reef=as.factor(Reef),
  Survey=as.factor(Survey),
  Site=as.factor(Site),
  T_ID = as.factor(T_ID),
  Depth = as.factor(Depth),
  Species_name = as.factor(Species_name))


monitoring<-read_csv("../0_data/Monitoring_species.csv") # all species from the UVU shallow surveys,a,b and functional group

monitoring <- monitoring %>% mutate(
  Family = as.factor(Family),
  Species = as.factor(Species),
  Species_name = as.factor(Species),
  `Functional group` = as.factor(`Functional group`)
)
```


# Unique species from ROV or BUV
```{R}
unique.rov<-rov.fish %>% 
  distinct(Species_name)

unique.bruv<-bruv.fish %>% 
  distinct(Species_name)

unique.rov<-unique.rov %>% filter(!str_detect(Species_name,"sp"))

unique.bruv<-unique.bruv %>% filter(!str_detect(Species_name,"sp"))

```

# Anti Match with monitoring species
```{r}
df<-anti_join(unique.rov, monitoring, by = "Species_name")
df<-df %>% distinct(Species_name)


df2<-anti_join(unique.bruv, monitoring, by = "Species_name")

df2<-df2 %>% distinct(Species_name)

```

# Get rid of anything only identified to "sp" in deep surveys
  NOW done in step above with unqiue.rov/bruv
 ```{r}

# df<-df %>% filter(!str_detect(Species_name,"sp"))
#df2<-df2 %>% filter(!str_detect(Species_name,"sp"))

# or regex (?!.*sp.*)^.*$
```

# Compare to species records data
```{r}
fish.records <-read_csv("../0_data/CSMP_fish_records.csv")  # 2018-2022 fish records from CSMP monitoring. includes species seen as well as those counted in surveys.

fish.records<- fish.records%>% rename(Species_name = Species)%>%
  mutate(Species_name = as.factor(Species_name))

```
# Anti Match with CSMP species records (includes having been matching with monitoring list)
```{r}
rov.newfish<-anti_join(df, fish.records, by = "Species_name")
rov.newfish<-rov.newfish %>% distinct(Species_name)


bruv.newfish<-anti_join(df2, fish.records, by = "Species_name")
bruv.newfish<-bruv.newfish %>% distinct(Species_name)

```



# Export to csv for checking
```{r}
write.csv(rov.newfish, file= "../2_outputs/newspecies_ROV.csv", row.names = FALSE)
write.csv(bruv.newfish, file= "../2_outputs/newspecies_BRUV.csv", row.names = FALSE)
```
